Add stack trace to Error object
Plain golang error object do not contains an stack trace, it is hard to
find where the error is.

Put stack into error hurt performace, but since it is an error, it is
rare to create an error object, and is acceptable to slow a bit on
error.

This idea, and some code from: https://github.com/go-errors/errors.

Since now error object is an heavy object, so put CausedBy as a field
instead of from lightweight interface, put to field simplifies code.